fix: labor-widget renamed to production-widget
