+++ client/apps/eternum-mobile/src/widgets/production-widget/index.ts
@@ -1,3 +1,3 @@
 export type { LaborBuilding, LaborBuildingProps, ResourceBalance } from "./model/types";
 export { LaborDrawer } from "./ui/labor-drawer";
-export { LaborWidget } from "./ui/labor-widget";
+export { ProductionWidget } from "./ui/production-widget";
