+++ client/apps/eternum-mobile/src/widgets/production-widget/model/types.ts
