+++ client/apps/eternum-mobile/src/widgets/production-widget/ui/labor-drawer.tsx
